Got the first async test working
Got an async test to run. Ran into some issues with Refit, this thing
seems hard to be made to work through a proxy like fiddler, also, if
there are some errors in the framework, everything fails miserably
without even hitting a breakpoint. Also, composing the Api using partial
Interfaces, seems like a no go for now, it fails the refit's code
generation. The good thing is that the tests feel so snappy.