Kind of getting to a somehow sane api for the Api
It looks like code generation is inevitable. C# doesn't have yet goodies
as decorators or mixins, probably there are a few OOP patterns out there
that can help with this situation, but most likely those will come with
the cost of intellisense. Since the ApiClient for testing will most
likely be a rigid thingie, probably generating it once in a while,
shouldn't be a bad idea.

Anyway, who knows, will play around with the Api, something may come to
mind.