Added the NNanomsg library
The DataManagerServer and DataManagerClient will require a method to
communicate with each other, since DataManagerServer might be running on
a separate machine in the network and be used concurrently by many
tests.

A communication mechanism is required, it was tempting to just do a REST
and JSON API, but decided to give nanomsg a try, NNanomsg is the C#
binding for it, so far it works good, I have high hopes for it, even
though it's a beta product. The bummer is that NNanomsg it's not on
nugget, will have to do this myself, also it will be an exercise to add
a nuget lib with native components, in case will have to make a rocks db
wrapper later.

So far so good, will clean everything up. Added the NNanomsg dlls to the
repository for now. Once the nuget package is done, will remove those.