Got serialization to work
Used the binary formatter to serialize an ApiObject and send it from the
DataManagerServer to the DataManagerClient and was successfully
deserialized and used by the test. Will have to do some more
refactoring, and maybe rethink the architecture but so far, so good.
Will remove the capnproto depdendency, hope the BinaryFormatter will be
fast enough for our needs.